# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Think Then Talk API
  description: The complete API specification for the Think Then Talk event platform.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /admin/registrations:
    get:
      tags:
        - Admin
      summary: List All Registrations (Admin)
      description: Retrieves a detailed list of all user registrations for the admin panel.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of registrations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserRegistrationDetails'
  '/admin/registrations/{registrationId}':
    get:
      tags:
        - Admin
      summary: Get Registration Details (Admin)
      description: Retrieves full details for a single registration.
      security:
        - bearerAuth: []
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Registration details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistrationDetails'
        '404':
          description: Registration not found.
    put:
      tags:
        - Admin
      summary: Update Registration (Admin)
      description: Manually updates a registration's status or the user's submitted form data.
      security:
        - bearerAuth: []
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Send only the fields you want to update. Both fields are optional.
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - PAID
                    - PENDING
                    - FAILED
                    - CANCELLED
                formData:
                  $ref: '#/components/schemas/RegistrationFormData'
      responses:
        '200':
          description: Registration updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistrationDetails'
        '404':
          description: Registration not found.
...
  /admin/registrations/export:
    get:
      tags:
        - Admin
      summary: 'Note: This endpoint is not currently called by the frontend; it is a planned/optional backend feature.'
      description: 'Note: This endpoint is not currently called by the frontend; it is a planned/optional backend feature.'
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum:
              - PAID
              - PENDING
              - FAILED
              - CANCELLED
      responses:
        '200':
          description: A CSV file of registrations.
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    TicketType:
      type: string
      enum:
        - EARLY_BIRD
        - STANDARD
        - SUPPORTER
    User:
      type: object
      required:
        - id
        - firstNameFa
        - lastNameFa
        - mobile
      properties:
        id:
          type: string
          description: The unique identifier for the user.
        firstNameFa:
          type: string
          description: User's first name in Persian.
        lastNameFa:
          type: string
          description: User's last name in Persian.
        firstNameEn:
          type: string
        lastNameEn:
          type: string
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - OTHER
        age:
          type: number
          format: int32
        educationLevel:
          type: string
        fieldOfStudy:
          type: string
        isEmployed:
          type: boolean
        jobTitle:
          type: string
        mobile:
          type: string
          description: The user's primary mobile number (unique).
        email:
          type: string
          format: email
        languageLevel:
          type: string
...
    RegistrationFormData:
      type: object
      description: The full set of data collected during user registration for an event.
      required:
        - firstNameFa
        - lastNameFa
        - mobile
        - isEmployed
        - acceptedRules
      properties:
        firstNameFa:
          type: string
        lastNameFa:
          type: string
        firstNameEn:
          type: string
        lastNameEn:
          type: string
        age:
          type: number
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - OTHER
        mobile:
          type: string
          description: User's mobile number used during registration.
        email:
          type: string
          format: email
        educationLevel:
          type: string
        fieldOfStudy:
          type: string
        isEmployed:
          type: boolean
        jobTitle:
          type: string
        languageLevel:
          type: string
        referralSource:
          type: string
        referrerName:
          type: string
        otherReferralSource:
          type: string
        acceptedRules:
          type: boolean
...
    UserRegistrationDetails:
      type: object
      required:
        - id
        - userId
        - eventId
        - paymentId
        - ticketType
        - status
        - createdAt
        - user
        - formData
        - event
        - payment
      properties:
        id:
          type: string
        userId:
          type: string
        eventId:
          type: string
        paymentId:
          type: string
        ticketType:
          $ref: '#/components/schemas/TicketType'
        status:
          type: string
          enum:
            - PAID
            - PENDING
            - FAILED
            - CANCELLED
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        formData:
          $ref: '#/components/schemas/RegistrationFormData'
        event:
          type: object
          properties:
            id:
              type: string
            title:
              type: string
            startDateTime:
              type: string
              format: date-time
        payment:
          type: object
          properties:
            id:
              type: string
            amount:
              type: number
            currency:
              type: string
              enum:
                - TOMAN
                - IRR
            status:
              type: string
              enum:
                - PENDING
                - SUCCESS
                - FAILED
            gatewayTransactionId:
              type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT